heba

+++ app/Http/Controllers/BranchController.php
@@ -14,6 +14,7 @@ class BranchController extends Controller
 {
     public function index()
     {
+        $this->authorize('viewAny' , [Branch::class]);
         $branches = Branch::all();
 
         $days = Day::all();
@@ -28,54 +29,31 @@ class BranchController extends Controller
         ]);
     }
 
-    public function showSpaces($id)
-    {
-        $branch = Branch::findOrFail($id);
-
-        // Retrieve the related spaces for the branch using the relationship (e.g., $branch->spaces)
-        $spaces = $branch->spaces;
-    
-        // You can also load any additional data you need for this view
-        $books = BookingRequest::all(); // Example: Retrieve booking requests
-    
-        return view('admin.branch.show_spaces', [
-            'branch' => $branch,
-            'spaces' => $spaces,
-            'books' => $books,
-            'space' => new Space(),
-            'branches' => Branch::all(),
-            'branchId' => $id,
-
-        ]);
-    }
 
     public function store(BranchRequest $request)
     {
+
+        $this->authorize('create' , [Branch::class]);
+
         $validatedData = $request->validated();
 
         $validatedData['user_id'] = Auth::user()->id;
 
         $branch = Branch::create($validatedData);
 
-        // Attach workdays to the newly created branch
         if ($request->has('work_days')) {
+
             $branch->workDays()->attach($request->input('work_days'));
         }
 
-        return back();
-        // return redirect()->route('branch.index')->with('success', __('Branch Created Successfully!'));
+        return back()->with('success', __('Branch Created Successfully!'));
     }
 
-    public function show(Branch $branch)
-    {
-
-        $days = Day::all();
-
-        return view('admin.branch.show', compact('branch', 'days'));
-    }
 
     public function update(BranchRequest $request, Branch $branch)
     {
+
+        $this->authorize('update', $branch);
 
         $validatedData = $request->validated();
 
@@ -89,11 +67,11 @@ class BranchController extends Controller
     public function destroy(Branch $branch)
     {
 
+        $this->authorize('delete', $branch);
+        
         $branch->workDays()->detach();
 
-        // Delete the branch
         $branch->delete();
-
 
         return redirect()->route('branch.index')->with('success', __('Branch Deleted Successfully!'));
     }
